finish delete user with alert & ajax

+++ public/scripts/doctor/deleteDoctor.js
@@ -1,25 +1,35 @@
-
 $(".deleteDoctor").click(function(){
-
-    console.log("received ID");
-    var id = $(this).data("id");
-    console.log(id);
-    var token = $("meta[name='csrf-token']").attr("content");;
-    console.log("received Token");
-    console.log(token);
     
-    $.ajax(
-    {
-        url: "/pharmacy/doctors/destroy/"+id,
-        type: 'DELETE',
-        data: {
-            "id": id,
-            "_token": token,
-        },
-        success: function (){
-            console.log("it Works");
-        }
-    });
+    if(!confirm("Do you really want to do this?")) {
+        console.log("It failed");
+        return false;
+      }
+      else{
+            console.log("received ID");
+            var id = $(this).data("id");
+            console.log(id);
+            var token = $("meta[name='csrf-token']").attr("content");;
+            console.log("received Token");
+            console.log(token);
+            var div=document.querySelector("#data"+id);
+            console.log(div);
+            $.ajax(
+            {
+                url: "/pharmacy/doctors/destroy/"+id,
+                type: 'DELETE',
+                data: {
+                    "id": id,
+                    "_token": token,
+                },
+                success: function (){
+                      div.remove();
 
-    console.log("It failed");
+                    Swal.fire(
+                        'Remind!',
+                        'Doctor deleted successfully!',
+                        'error'
+                    )
+                }
+            });
+      }
 });
